rename cart to checkout and add brief component
