--- src/components/Cart/index.js
@@ -1,1 +1,0 @@
-export { default as Cart } from "./Cart";
