+++ src/components/Checkout/Brief/Brief.jsx
@@ -1,0 +1,13 @@
+import "./Brief.css";
+
+const Brief = ({ getTotalPrice, clearCart, finishPurchase }) => {
+  return (
+    <div>
+      <p>Total: ${getTotalPrice()}</p>
+      <button onClick={() => clearCart()}>Clear cart</button>
+      <button onClick={() => finishPurchase()}>Finish purchase</button>
+    </div>
+  );
+};
+
+export default Brief;
