+++ src/components/Checkout/Brief/index.js
@@ -1,0 +1,1 @@
+export { default as Brief } from "./Brief";
