+++ src/components/Checkout/CartItem/CartItem.css
@@ -1,13 +1,3 @@
-.cartContainer {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  margin: 0 auto;
-  width: 100%;
-  height: 100%;
-  padding: 0 1rem;
-}
 .cart {
   display: flex;
   flex-direction: row;
