+++ src/components/Checkout/CartItem/index.js
@@ -1,0 +1,1 @@
+export { default as CartItem } from "./CartItem";
