+++ src/components/Checkout/index.js
@@ -1,0 +1,1 @@
+export { default as Checkout } from "./Checkout";
