+++ src/components/CheckoutContainer/CheckoutContainer.jsx
@@ -1,9 +1,9 @@
 import { Layout } from "../Layout";
-import { Checkout } from "../Checkout";
+import { Cart } from "../Checkout";
 import { useContext } from "react";
 import { CartContext } from "../../context/CartContext";
 
-const CartContainer = () => {
+const CheckoutContainer = () => {
   const { cart, removeFromCart, clearCart, finishPurchase, getTotalPrice } = useContext(CartContext);
 
   return (
@@ -12,10 +12,10 @@ const CartContainer = () => {
         <div className="welcome">
           <h1 className="greeting">Shopping Cart</h1>
         </div>
-        <Checkout cart={cart} removeFromCart={removeFromCart} clearCart={clearCart} finishPurchase={finishPurchase} getTotalPrice={getTotalPrice} />
+        <Cart cart={cart} removeFromCart={removeFromCart} clearCart={clearCart} finishPurchase={finishPurchase} getTotalPrice={getTotalPrice} />
       </div>
     </Layout>
   );
 };
 
-export default CartContainer;
+export default CheckoutContainer;
